Fix callback URL to something "supported" by Xero
The Xero OAuth service requires a fragment in the URL. The fragment can
be anything, it just can't be empty. So http://127.0.0.1:3000 doesn't
work - the Xero firewall deems it some kind of remote file inclusion
attack and blocks it, returning an HTTP 403.
http://172.0.0.1:3000/something does work. Our mock server
Test::HTTP::MockServer::Once does not care about fragments, it returns
the same content no matter the fragment, so that's fine. Took me quite
some time with Xero support to figure this out.